docs: separate chip & chipgroup stories
